requirements.txt
python-telegram-bot
ibm_watson
python-dotenv
pycbrf